Book mybatis 연결 big decimal > int

+++ 02_java/book/src/book/exception/NotFoundException.java
@@ -19,6 +19,6 @@ public class NotFoundException extends Exception {
 	
 	// 중복정의 생성자
 	public NotFoundException(String type, Book book) {
-		super(String.format("%s:[bookSeq:%d] 도서 정보가 존재하지 않습니다.", type, book.getBookSeq().intValue()));
+		super(String.format("%s:[bookSeq:%d] 도서 정보가 존재하지 않습니다.", type, book.getBookSeq()));
 	}
 }
